pyproject.toml
[project]
name = "customer-support-agent"
version = "0.1.0"
description = "Customer support triage agent using LangGraph and Tavily"
requires-python = ">=3.11"
dependencies = [
    "langgraph",
    "langchain-core",
    "openai==1.52.2",
    "faiss-cpu",
    "numpy",
    "tavily-python",
    "python-dotenv",
    "httpx<0.28.0",
    "jupyter>=1.1.1",
    "ipykernel>=6.30.1",
    "langchain-upstage>=0.5.0",
    "langchain>=0.3.27",
    "langchain-community",
    "langgraph-cli[inmem]>=0.4.2",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
]
